Add --sei (same start-end-identical)
Try to filter out diffs where the route never reaches the same
destination. This can happen when the random location is far of
the road network and the snap-on road is not in the graph anymore
due to profile changes. This may also happen on the edges of
your area/pbf file.